requirements.txt
streamlit>=1.33.0
python-dotenv>=1.0.1
langchain-core>=0.2.0
langchain-google-genai>=0.1.7
google-api-core>=2.19.0
google-generativeai>=0.5.2
